feat: add extents
