+++ src/components/collections.tsx
@@ -1,7 +1,6 @@
-import { Card, Heading, Link, Stack, Text } from "@chakra-ui/react";
-import { MarkdownHooks } from "react-markdown";
+import { Heading, Stack } from "@chakra-ui/react";
 import type { StacCollection } from "stac-ts";
-import useStacMap from "../hooks/stac-map";
+import { CollectionCard } from "./collection";
 
 export default function Collections({
   collections,
@@ -20,27 +19,5 @@ export default function Collections({
         ))}
       </Stack>
     </Stack>
-  );
-}
-
-function CollectionCard({ collection }: { collection: StacCollection }) {
-  const { setHref } = useStacMap();
-  const selfHref = collection.links.find((link) => link.rel === "self")?.href;
-
-  return (
-    <Card.Root size={"sm"}>
-      <Card.Body>
-        <Card.Title>
-          <Link onClick={() => selfHref && setHref(selfHref)}>
-            {collection.title || collection.id}
-          </Link>
-        </Card.Title>
-        <Card.Description>
-          <Text lineClamp={2}>
-            <MarkdownHooks>{collection.description}</MarkdownHooks>
-          </Text>
-        </Card.Description>
-      </Card.Body>
-    </Card.Root>
   );
 }
